러닝 모집 게시글 작성 UI 완성

+++ presentation/src/main/kotlin/team/applemango/runnerbe/activity/MainActivity.kt
@@ -19,7 +19,8 @@ import dagger.hilt.android.AndroidEntryPoint
 import team.applemango.runnerbe.R
 import team.applemango.runnerbe.databinding.ActivityMainBinding
 import team.applemango.runnerbe.shared.constant.BottomNavigationBarHeight
-import team.applemango.runnerbe.shared.util.extension.dp
+import team.applemango.runnerbe.shared.domain.unit.dp
+import team.applemango.runnerbe.shared.domain.unit.toInt
 
 @AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
@@ -43,7 +44,7 @@ class MainActivity : AppCompatActivity() {
             supportFragmentManager.findFragmentById(R.id.fcv_main) as NavHostFragment
         val navController = navHostFragment.navController
         binding.bnvMenu.updateLayoutParams {
-            height = BottomNavigationBarHeight.dp
+            height = BottomNavigationBarHeight.dp.toInt()
         }
         binding.bnvMenu.setupWithNavController(navController)
     }
